fix: move basic sample to typescript
