+++ examples/basic/android/app/build.gradle
@@ -66,7 +66,7 @@ import com.android.build.OutputFile
  */
 
 project.ext.react = [
-    entryFile: "index.android.js",
+    entryFile: "index.js",
     enableHermes: false,
 ]
 apply from: "../../node_modules/react-native/react.gradle"
