+++ examples/basic/index.js
@@ -1,0 +1,5 @@
+import VideoPlayer from './VideoPlayer';
+
+import { AppRegistry } from 'react-native';
+
+AppRegistry.registerComponent('VideoPlayer', () => VideoPlayer);
