+++ examples/basic/ios/main.jsbundle
@@ -1,5 +1,5 @@
 // Offline JS
 // To re-generate the offline bundle, run this from root of your project
-// $ curl 'http://localhost:8081/index.ios.bundle?dev=false&minify=true' -o iOS/main.jsbundle
+// $ curl 'http://localhost:8081/index.bundle?dev=false&minify=true' -o iOS/main.jsbundle
 
 throw new Error('Offline JS file is empty. See iOS/main.jsbundle for instructions');
