+++ examples/basic/package.json
@@ -18,6 +18,8 @@
     "@babel/core": "^7.6.0",
     "@babel/runtime": "^7.6.0",
     "@react-native-community/eslint-config": "^0.0.5",
+    "@types/react": "^18.0.6",
+    "@types/react-native": "^0.67.6",
     "babel-jest": "^24.9.0",
     "eslint": "^6.4.0",
     "jest": "^24.9.0",
